MET-20: combine the "/deploy/up" and "/deploy/[tag]/status" endpoints
